Added high score functionality
Changes to score.c and project.c to manage high scores. Also touched up
current store display to match values and maximum length.

+++ terminalio.h
@@ -87,6 +87,6 @@ void draw_horizontal_line(int8_t y, int8_t startx, int8_t endx);
 void draw_vertical_line(int8_t x, int8_t starty, int8_t endy);
 
 //display the current score
-void display_score(uint16_t);
+void display_score(uint32_t);
 
 #endif /* TERMINAL_IO_H */
